requirements.txt
flask-security-too[common,mfa,fsqla]
zxcvbn
pandas
mysql-connector-python
python-dotenv
flask-cors
sqlalchemy<2.0
